The marching cubes algorithm is no longer called on a material change.
i.e. No longer called unless needed.

This has been done with various if statements and could be improved.

---

surface no longer holds the colour and opacity -> only the model does

Job now takes a Double instead of a double to allow for a NULL parameter.
- MarchingCubesModel has been edited to follow this.

IsoHandler passes null for the isovalue and cube size parameters in the job if either the colour of opacity are changed.