Masking tool bug fix
As the FloatSpinner class now, by default, declares a length of 3 inherently all boxes in the masking tool were limited to entry of 3 digit numbers, with a precision of one (i.e. +\- 99.9). Now the length of Integer.MAX_VALUE is used instead of simply 3, however, it is known that this fix for this makes the spinner box much larger than is required but a fix resolving this does not appear forthcoming.