DAQ-826 Fix old region draw events are not cleared
If multiple region draw events are triggered but not completed the
listeners on the eventListeners on the regionOverlayArea could
accumulate causing incorrect region types to be drawn.

This fix could probably be improved by better encapsulation of the
listener management or by creating a new layer when a draw is needed.