# WARNING: head commit changed in the meantime
SCI-7223 fix single frame processing

The operation context is setup by the first file and if that file has only one frame then the code left that frame dimension as a "slice". This fix ensures that slice is switched to a range so that all frames are used.

Note, this may not be suffice for scans with two or more dimensions...