Improve management of state in DataVis
The aim of this refactor was to improve the thread safety and hopefully
make the state a bit simpler, as well as tidying some APIs.

The main classes changed are the FileController and PlotController, but
as these are the core of the perspective there are many associated changes.

Its a bit of a large change, and may need a few more improvements, but
at this point everything seems to work.

Change-Id: Ic609e745fe0cac9ef042aa893b8e034527308fbc