Fix image inversion bug for I21-605
The y axis is inconsistent with its range when an image trace is created. Now setInverted will set the range accordingly to be consistent. Also, setting range will override with inversion state.

Inverting an axis via the config dialog will change the image origin and update the drop-down menu selection

Change-Id: I83f71e40bb1b8ad2390827504296a04ca73c659b