SCI-9574 Show error when saving to read only directory
If an invalid file location is chosen when saving a screenshot, an
exception is raised and the chosen file name is not returned. If there
was no previous filename to fall back to, a NPE prevented the error
message being displayed. If a previous filename was available, the error
message would be shown but would relate to the wrong file path.

Don't try to include the filename in the error and instead rely on the
included exception cause to include enough information (it does include
the filename for read-only errors).

Change-Id: I32c6d0acb0cfe9584317c935139b2fdd0f9d7322