Cache file state at load for uninitialised files
I07 want the filecontroller to load new files showing the same dataset
as the last plotted data, but they also want that data to be deselected
when a new file is loaded. This would work fine but when datavis loads
the file, initially no datasets are populated (cant read file) so state
is cleared and not propagated to new file. Caching state fixes this.

Change-Id: I70fe37cd12fb2f0eb02dae760fb8cb40dce06d5d