Prevent IndexOutOfBoundExceptions when configuring traces
The trace index was being used to select the traces tab from the dialog.
If there were more than three traces, this led to an exception and the
dialog was not shown.

Instead, use the parent of the trace composite as the traces tab root
composite when setting the top level control to display.

Change-Id: Ied11b09a056b16359696e9ec7e81bd660c794f5f